Moved Place and Region into HttpAPI folder and namespace.
