+++ src/UnitySL/Assets/Scripts/HttpAPI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1987044eb5d03094291746be2f7e992f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
